# app/static/swagger.json
{"paths": {"/api/v1/auth/register": {"post": {"tags": ["Authentication"], "summary": "\u0420\u0435\u0433\u0438\u0441\u0442\u0440\u0430\u0446\u0438\u044f \u043d\u043e\u0432\u043e\u0433\u043e \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f", "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "email": {"type": "string", "format": "email"}, "password": {"type": "string", "format": "password"}}, "required": ["username", "email", "password"]}}}}, "responses": {"201": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0437\u0430\u0440\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u043e\u0432\u0430\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445"}}}}, "/api/v1/auth/login": {"post": {"tags": ["Authentication"], "summary": "\u0412\u0445\u043e\u0434 \u0432 \u0441\u0438\u0441\u0442\u0435\u043c\u0443", "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string", "format": "password"}}, "required": ["username", "password"]}}}}, "responses": {"200": {"description": "\u0423\u0441\u043f\u0435\u0448\u043d\u044b\u0439 \u0432\u0445\u043e\u0434", "content": {"application/json": {"schema": {"type": "object", "properties": {"access_token": {"type": "string"}, "user": {"$ref": "#/components/schemas/User"}}}}}}, "401": {"description": "\u041d\u0435\u0432\u0435\u0440\u043d\u044b\u0435 \u0443\u0447\u0435\u0442\u043d\u044b\u0435 \u0434\u0430\u043d\u043d\u044b\u0435"}}}}, "/api/v1/users": {"get": {"tags": ["Users"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0441\u043f\u0438\u0441\u043a\u0430 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u0435\u0439 (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "responses": {"200": {"description": "\u0421\u043f\u0438\u0441\u043e\u043a \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u0435\u0439", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/User"}}}}}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}}}, "post": {"tags": ["Users"], "summary": "\u0421\u043e\u0437\u0434\u0430\u043d\u0438\u0435 \u043d\u043e\u0432\u043e\u0433\u043e \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "email": {"type": "string", "format": "email"}, "password": {"type": "string", "format": "password"}, "role": {"type": "string", "enum": ["admin", "respondent"]}}, "required": ["username", "email", "password", "role"]}}}}, "responses": {"201": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0441\u043e\u0437\u0434\u0430\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}}}}, "/api/v1/users/{user_id}": {"get": {"tags": ["Users"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0434\u0430\u043d\u043d\u044b\u0445 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f", "security": [{"BearerAuth": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "\u0414\u0430\u043d\u043d\u044b\u0435 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}, "put": {"tags": ["Users"], "summary": "\u041e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u0430\u043d\u043d\u044b\u0445 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["admin", "respondent"]}}}}}}, "responses": {"200": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}, "delete": {"tags": ["Users"], "summary": "\u0423\u0434\u0430\u043b\u0435\u043d\u0438\u0435 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"204": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0443\u0434\u0430\u043b\u0435\u043d"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}}, "/api/v1/restaurants": {"get": {"tags": ["Restaurants"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0441\u043f\u0438\u0441\u043a\u0430 \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u043e\u0432", "responses": {"200": {"description": "\u0421\u043f\u0438\u0441\u043e\u043a \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u043e\u0432", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Restaurant"}}}}}}}, "post": {"tags": ["Restaurants"], "summary": "\u0421\u043e\u0437\u0434\u0430\u043d\u0438\u0435 \u043d\u043e\u0432\u043e\u0433\u043e \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430 (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"name": {"type": "string"}, "address": {"type": "string"}, "description": {"type": "string"}}, "required": ["name"]}}}}, "responses": {"201": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0441\u043e\u0437\u0434\u0430\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Restaurant"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}}}}, "/api/v1/restaurants/{restaurant_id}": {"get": {"tags": ["Restaurants"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0434\u0430\u043d\u043d\u044b\u0445 \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430", "parameters": [{"name": "restaurant_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "\u0414\u0430\u043d\u043d\u044b\u0435 \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Restaurant"}}}}, "404": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}, "put": {"tags": ["Restaurants"], "summary": "\u041e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0434\u0430\u043d\u043d\u044b\u0445 \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430 (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "parameters": [{"name": "restaurant_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"name": {"type": "string"}, "address": {"type": "string"}, "description": {"type": "string"}}}}}}, "responses": {"200": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u043e\u0431\u043d\u043e\u0432\u043b\u0435\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Restaurant"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}, "delete": {"tags": ["Restaurants"], "summary": "\u0423\u0434\u0430\u043b\u0435\u043d\u0438\u0435 \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430 (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "parameters": [{"name": "restaurant_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"204": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0443\u0434\u0430\u043b\u0435\u043d"}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}}, "/api/v1/restaurants/report": {"get": {"tags": ["Reports"], "summary": "\u0412\u044b\u0433\u0440\u0443\u0437\u043a\u0430 \u0441\u0432\u043e\u0434\u043d\u043e\u0433\u043e \u043e\u0442\u0447\u0435\u0442\u0430 \u043f\u043e \u0432\u0441\u0435\u043c \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430\u043c (\u0442\u043e\u043b\u044c\u043a\u043e \u0434\u043b\u044f \u0430\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u043e\u0440\u043e\u0432)", "security": [{"BearerAuth": []}], "responses": {"200": {"description": "\u0421\u0432\u043e\u0434\u043d\u044b\u0439 \u043e\u0442\u0447\u0435\u0442 \u0432 \u0444\u043e\u0440\u043c\u0430\u0442\u0435 CSV", "content": {"text/csv": {"schema": {"type": "string", "format": "binary"}}}}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}}}}, "/api/v1/reviews": {"get": {"tags": ["Reviews"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0441\u043f\u0438\u0441\u043a\u0430 \u043e\u0442\u0437\u044b\u0432\u043e\u0432", "security": [{"BearerAuth": []}], "parameters": [{"name": "restaurant_id", "in": "query", "schema": {"type": "integer"}, "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e ID \u0440\u0435\u0441\u0442\u043e\u0440\u0430\u043d\u0430"}, {"name": "user_id", "in": "query", "schema": {"type": "integer"}, "description": "\u0424\u0438\u043b\u044c\u0442\u0440 \u043f\u043e ID \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f"}], "responses": {"200": {"description": "\u0421\u043f\u0438\u0441\u043e\u043a \u043e\u0442\u0437\u044b\u0432\u043e\u0432", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Review"}}}}}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}}}, "post": {"tags": ["Reviews"], "summary": "\u0421\u043e\u0437\u0434\u0430\u043d\u0438\u0435 \u043d\u043e\u0432\u043e\u0433\u043e \u043e\u0442\u0437\u044b\u0432\u0430", "security": [{"BearerAuth": []}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"restaurant_id": {"type": "integer"}, "food_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "drinks_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "overall_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "comment": {"type": "string"}}, "required": ["restaurant_id", "food_rating", "drinks_rating", "overall_rating"]}}}}, "responses": {"201": {"description": "\u041e\u0442\u0437\u044b\u0432 \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0441\u043e\u0437\u0434\u0430\u043d", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Review"}}}}, "400": {"description": "\u041e\u0448\u0438\u0431\u043a\u0430 \u0432\u0430\u043b\u0438\u0434\u0430\u0446\u0438\u0438 \u0434\u0430\u043d\u043d\u044b\u0445 \u0438\u043b\u0438 \u043e\u0442\u0437\u044b\u0432 \u0443\u0436\u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442"}, "401": {"description": "\u0422\u0440\u0435\u0431\u0443\u0435\u0442\u0441\u044f \u0430\u0443\u0442\u0435\u043d\u0442\u0438\u0444\u0438\u043a\u0430\u0446\u0438\u044f"}, "404": {"description": "\u0420\u0435\u0441\u0442\u043e\u0440\u0430\u043d \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}}, "/api/v1/reviews/{review_id}": {"get": {"tags": ["Reviews"], "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0434\u0430\u043d\u043d\u044b\u0445 \u043e\u0442\u0437\u044b\u0432\u0430", "security": [{"BearerAuth": []}], "parameters": [{"name": "review_id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "\u0414\u0430\u043d\u043d\u044b\u0435 \u043e\u0442\u0437\u044b\u0432\u0430", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Review"}}}}, "403": {"description": "\u0414\u043e\u0441\u0442\u0443\u043f \u0437\u0430\u043f\u0440\u0435\u0449\u0435\u043d"}, "404": {"description": "\u041e\u0442\u0437\u044b\u0432 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d"}}}}}, "info": {"title": "Restaurant Reviews API", "version": "1.0.0"}, "openapi": "3.0.2", "components": {"schemas": {"User": {"type": "object", "properties": {"id": {"type": "integer"}, "username": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["admin", "respondent"]}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}}, "Restaurant": {"type": "object", "properties": {"id": {"type": "integer"}, "name": {"type": "string"}, "address": {"type": "string"}, "description": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}}, "Review": {"type": "object", "properties": {"id": {"type": "integer"}, "restaurant_id": {"type": "integer"}, "user_id": {"type": "integer"}, "food_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "drinks_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "overall_rating": {"type": "integer", "minimum": 1, "maximum": 5}, "comment": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}}}, "securitySchemes": {"BearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}}}}